\name{designGG-internal}

\alias{nSlidesANDnTuple}

\title{Internal designGG functions}

\description{
  Internal designGG functions. These are generally not to be called by the user.
}

\usage{
  TODO
}


\author{ 
	Yang Li \email{yang.li@rug.nl}
	Maintainer: Yang Li \email{yang.li@rug.nl}
}

\keyword{internal}